CMakeLists simplification, moves sources for more coherency, set a way to add first test
