+++ MMT/CMakeLists.txt
@@ -26,32 +26,29 @@ find_package(Qt${QT_VERSION_MAJOR} REQUIRED COMPONENTS Widgets LinguistTools Net
 
 set(TS_FILES MMT_en_US.ts)
 
-file(GLOB COMMON_GRAPHICS_HEADERS "CommonGraphics/*.h")
 file(GLOB COMMON_GRAPHICS_SOURCES "CommonGraphics/*.cpp")
-file(GLOB DIAGRAM_SCENES_HEADERS "DiagramScenes/*.h")
 file(GLOB DIAGRAM_SCENES_SOURCES "DiagramScenes/*.cpp")
-file(GLOB GRAPHICS_ITEMS_HEADERS "GraphicsItems/*.h")
 file(GLOB GRAPHICS_ITEMS_SOURCES "GraphicsItems/*.cpp")
-file(GLOB TOOLBOX_SCENES_HEADERS "ToolBoxScenes/*.h")
 file(GLOB TOOLBOX_SCENES_SOURCES "ToolBoxScenes/*.cpp")
-file(GLOB CONFIGURATION_CONTEXTS_HEADERS "ConfigurationContexts/*.h")
 file(GLOB CONFIGURATION_CONTEXTS_SOURCES "ConfigurationContexts/*.cpp")
-file(GLOB ENGINE_HEADERS "Engine/*.h")
 file(GLOB ENGINE_SOURCES "Engine/*.cpp")
-file(GLOB SHARING_HEADERS "Sharing/*.h")
 file(GLOB SHARING_SOURCES "Sharing/*.cpp")
+
+set(PROJECT_SOURCES_WITHOUT_MAIN
+    ${COMMON_GRAPHICS_SOURCES}
+    ${DIAGRAM_SCENES_SOURCES}
+    ${GRAPHICS_ITEMS_SOURCES}
+    ${TOOLBOX_SCENES_SOURCES}
+    ${CONFIGURATION_CONTEXTS_SOURCES}
+    ${ENGINE_SOURCES}
+    ${SHARING_SOURCES}
+)
 
 set(PROJECT_SOURCES
     main.cpp
     MeMoToApplication.h MeMoToApplication.cpp MeMoToThemeDefinition.h
     MainWindow.h MainWindow.cpp
-    ${COMMON_GRAPHICS_HEADERS} ${COMMON_GRAPHICS_SOURCES}
-    ${DIAGRAM_SCENES_HEADERS} ${DIAGRAM_SCENES_SOURCES}
-    ${GRAPHICS_ITEMS_HEADERS} ${GRAPHICS_ITEMS_SOURCES}
-    ${TOOLBOX_SCENES_HEADERS} ${TOOLBOX_SCENES_SOURCES}
-    ${CONFIGURATION_CONTEXTS_HEADERS} ${CONFIGURATION_CONTEXTS_SOURCES}
-    ${ENGINE_HEADERS} ${ENGINE_SOURCES}
-    ${SHARING_HEADERS} ${SHARING_SOURCES}
+    ${PROJECT_SOURCES_WITHOUT_MAIN}
     ${TS_FILES}
 )
