+++ MMT/GraphicsItems/EntryStateGraphicsItem.h
@@ -6,7 +6,7 @@
 #include <Engine/I_Selectable.h>
 #include <Engine/I_Connectable.h>
 
-#include <GraphicsItems/HandleGraphicsItem.h>
+#include <CommonGraphics/HandleGraphicsItem.h>
 
 class EntryStateGraphicsItem:
         public I_Connectable
