+++ MMT/tests/CommonGraphics/CMakeLists.txt
@@ -10,19 +10,31 @@ file(GLOB headers "*.h")
 find_package(QT NAMES Qt6 Qt5 REQUIRED COMPONENTS Core Gui Widgets LinguistTools Svg Qt6Test)
 find_package(Qt${QT_VERSION_MAJOR} REQUIRED COMPONENTS Core Gui Widgets LinguistTools Network WebSockets Svg Test)
 
-foreach(header ${headers})
-    get_filename_component(executable_name "${header}" NAME_WE)
-    add_executable(${executable_name} ${header} ${executable_name}.cpp)
-    add_test(${executable_name} ${executable_name})
-
-    target_link_libraries(${executable_name}
-      PRIVATE
-        Qt${QT_VERSION_MAJOR}::Widgets
-	Qt${QT_VERSION_MAJOR}::Test
-    )
-
-    if(QT_VERSION_MAJOR EQUAL 6)
-      qt_finalize_executable(${executable_name})
-    endif()
-
-endforeach()
+# foreach(header ${headers})
+#     get_filename_component(executable_name "${header}" NAME_WE)
+#     add_executable(
+#         ${executable_name} ${header} ${executable_name}.cpp
+#         ${COMMON_GRAPHICS_SOURCES}
+#         ${ENGINE_SOURCES}
+#         ${SHARING_SOURCES}
+#         ${CONFIGURATION_CONTEXTS_SOURCES}
+#     )
+#     add_test(${executable_name} ${executable_name})
+#
+#     target_include_directories(${executable_name} PRIVATE ${CMAKE_SOURCE_DIR})
+#
+#     target_link_libraries(${executable_name}
+#       PRIVATE
+#         Qt${QT_VERSION_MAJOR}::Widgets
+#         Qt${QT_VERSION_MAJOR}::Network
+#         Qt${QT_VERSION_MAJOR}::WebSockets
+#         Qt${QT_VERSION_MAJOR}::Svg
+#         Qt${QT_VERSION_MAJOR}::Test
+#     )
+#
+#     if(QT_VERSION_MAJOR EQUAL 6)
+#       qt_finalize_executable(${executable_name})
+#     endif()
+#
+# endforeach()
+#
