+++ MMT/tests/CommonGraphics/TestConfigWidget.cpp
@@ -1,8 +1,12 @@
 #include <TestConfigWidget.h>
 
+#include <CommonGraphics/ConfigWidget.h>
+
 QTEST_MAIN(TestConfigWidget)
 
 void TestConfigWidget::open()
 {
     QVERIFY(true);
+
+    ConfigWidget::getInstance();
 }
